+++ envelope/__init__.py
@@ -1,3 +1,4 @@
 from .envelope import Envelope
 
+Envelope.default = Envelope()
 __all__ = ["Envelope"]