+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## 1.1 (unreleased)
+- fix: default object
+- better EML files loading
 
 ## 1.0.0 (2020-05-28)
 - CHANGED: